# Intrinsics support for WidenFourAsciiBytesToUtf16AndWriteToBuffer