--- a/src/libraries/System.Utf8String.Experimental/src/System/Runtime/Intrinsics/Intrinsics.Shims.cs
+++ b/src/libraries/System.Utf8String.Experimental/src/System/Runtime/Intrinsics/Intrinsics.Shims.cs
@@ -27,6 +27,8 @@ namespace System.Runtime.Intrinsics
         public static Vector128<ulong> AsUInt64<T>(this Vector128<T> vector) where T : struct => throw new PlatformNotSupportedException();
         public static T GetElement<T>(this Vector128<T> vector, int index) where T : struct => throw new PlatformNotSupportedException();
         public static T ToScalar<T>(this Vector64<T> vector) where T : struct => throw new PlatformNotSupportedException();
+        public static unsafe Vector128<ulong> CreateScalar(ulong value) => throw new PlatformNotSupportedException();
+        public static T ToScalar<T>(this Vector128<T> vector) where T : struct => throw new PlatformNotSupportedException();
     }
     internal readonly struct Vector128<T>
         where T : struct
@@ -128,27 +130,13 @@ namespace System.Runtime.Intrinsics.X86
     }
 }
 
-namespace System.Runtime.Intrinsics.Arm
+namespace System.Runtime.CompilerServices
 {
-    internal abstract class ArmBase
+    // Calls to methods or references to fields marked with this attribute may be replaced at
+    // some call sites with jit intrinsic expansions.
+    // Types marked with this attribute may be specially treated by the runtime/compiler.
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Method | AttributeTargets.Constructor | AttributeTargets.Field, Inherited = false)]
+    internal sealed class IntrinsicAttribute : Attribute
     {
-        public abstract class Arm64
-        {
-            public const bool IsSupported = false;
-            public static int LeadingZeroCount(ulong value) => throw new PlatformNotSupportedException();
-            public static uint ReverseElementBits(ulong value) => throw new PlatformNotSupportedException();
-        }
-        public const bool IsSupported = false;
-        public static int LeadingZeroCount(uint value) => throw new PlatformNotSupportedException();
-        public static uint ReverseElementBits(uint value) => throw new PlatformNotSupportedException();
-    }
-
-    internal abstract class AdvSimd : ArmBase
-    {
-        public new abstract class Arm64 : ArmBase.Arm64
-        {
-            public static Vector64<byte> AddAcross(Vector64<byte> value) => throw new PlatformNotSupportedException();
-        }
-        public static Vector64<byte> PopCount(Vector64<byte> value) => throw new PlatformNotSupportedException();
     }
 }