Implements 1 API from https://github.com/dotnet/runtime/issues/35034