@GrabYourPitchforks 2020-07-01
Control flow LGTM, but I don't know enough about advsimd intrinsics to give this a proper review. @tannergooding, thoughts?
I really need to sit down with an arm64 machine for a few days and tinker. :)