@pgovind 2020-07-09
Thanks for the stamp Tanner. FWIW, here are the micro benchmarks I ran on my machine (Yeeesh, that took an unreasonable amount of time to set up):

D:\repos\performance\src\tools\ResultsComparer>dotnet run --base "..\..\..\..\before_arm64" --diff "..\..\..\..\after_arm64" --threshold 0.01%
summary:
better: 1, geomean: 2.057
total diff: 1

No Slower results for the provided threshold = 0.01% and noise filter = 0.3ns.

| Faster                                                                          | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Experimental.Perf.ToUtf16(expected: "This is a big string of words. |      2.06 |          1207.03 |           586.81 |         |